review comment fixes
