--- WalletWasabi/Discoverability/NostrCoordinatorConfiguration.cs
@@ -1,5 +1,0 @@
-using NBitcoin;
-
-namespace WalletWasabi.Discoverability;
-
-public record NostrCoordinatorConfiguration(string Description, Uri Uri, Network Network);
