+++ WalletWasabi/Discoverability/NostrCoordinatorPublisher.cs
@@ -15,26 +15,48 @@ public class NostrCoordinatorPublisher : PeriodicRunner
 {
 	public NostrCoordinatorPublisher(TimeSpan period, AnnouncerConfig config, Network network) : base(period)
 	{
+		Config = config;
+		Network = network;
 		Key = config.Key.FromNIP19Nsec();
-		CoordinatorConfiguration = new NostrCoordinatorConfiguration(config.CoordinatorDescription, new Uri(config.CoordinatorUri), network);
-		Client = NostrExtensions.Create(config.RelayUris.Select(x => new Uri(x)).ToArray(), (EndPoint?)null);
 	}
 
-	private INostrClient Client { get; }
-
-	private NostrCoordinatorConfiguration CoordinatorConfiguration { get; }
-
+	private AnnouncerConfig Config { get; }
+	private Network Network { get; }
 	private ECPrivKey Key { get; }
 
 	protected override async Task ActionAsync(CancellationToken cancel)
 	{
-		var discoveryEvent = await CoordinatorConfiguration.CreateCoordinatorDiscoveryEventAsync(Key).ConfigureAwait(false);
+		using var client = NostrExtensions.Create(Config.RelayUris.Select(x => new Uri(x)).ToArray(), (EndPoint?)null);
+		var discoveryEvent = await CreateCoordinatorDiscoveryEventAsync(Config, Key, Network).ConfigureAwait(false);
 
 		using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(10));
 		using var linkedCts = CancellationTokenSource.CreateLinkedTokenSource(cts.Token, cancel);
-		await Client.PublishAsync([discoveryEvent], linkedCts.Token).ConfigureAwait(false);
+		await client.PublishAsync([discoveryEvent], linkedCts.Token).ConfigureAwait(false);
 
 		Logger.LogInfo("Coordinator has been successfully published on Nostr.");
+	}
+
+	private async Task<NostrEvent> CreateCoordinatorDiscoveryEventAsync(AnnouncerConfig config, ECPrivKey key, Network network)
+	{
+		var evt = new NostrEvent()
+		{
+			Kind = 15750,
+			Content = config.CoordinatorDescription,
+			Tags =
+			[
+				CreateTag("endpoint", config.CoordinatorUri),
+				CreateTag("type", "wabisabi"),
+				CreateTag("network", network.ChainName.ToString().ToLower())
+			]
+		};
+
+		await evt.ComputeIdAndSignAsync(key).ConfigureAwait(false);
+		return evt;
+	}
+
+	private static NostrEventTag CreateTag(string tagIdentifier, string data)
+	{
+		return new() { TagIdentifier = tagIdentifier, Data = [data] };
 	}
 
 	public override void Dispose()
